Win check
Finished the winCheck function by concatenating a
new array of the board, an array of the board
columns, and an array of the board diagonals and
then iterating over each of the subarrays and
checking for each value being equal to either
players token. If neither player has won it
returns undefined, which I can check for in my
final render function. Or after each pick
potentially.