Board rendering and listeners
I've completed the listeners for the board cells.
I had to add them to the render function so that
each listener is created when the board renders.

I also figured our how to pass the created cells,
which I gave an ID of their board array indeces,
into the pick function, by splitting the id string
into an array, converting each to a number and
then passing each individually to the pick
function.