Board reset and wincheck multiplication
I have now added a function to reset the board to
allow for more games. I have also remade the
wincheck part of the pick function so that it
doesn't have to execute the wincheck function
multiple times every time, it just does it once
every time a cell is picked.